♻️ [RUMF-1267] remove dependency for `display`
