+++ packages/core/src/domain/console/consoleObservable.ts
@@ -3,16 +3,7 @@ import { computeStackTrace } from '../tracekit'
 import { createHandlingStack, formatErrorMessage, toStackTraceString } from '../../tools/error'
 import { mergeObservables, Observable } from '../../tools/observable'
 import { find, jsonStringify } from '../../tools/utils'
-
-export const ConsoleApiName = {
-  log: 'log',
-  debug: 'debug',
-  info: 'info',
-  warn: 'warn',
-  error: 'error',
-} as const
-
-export type ConsoleApiName = typeof ConsoleApiName[keyof typeof ConsoleApiName]
+import { ConsoleApiName } from '../../tools/display'
 
 export interface ConsoleLog {
   message: string
